+++ src/lib.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+use doc_comment;
+
 pub use context::HeifContext;
 pub use encoder::{EncoderParametersTypes, EncodingOptions};
 pub use enums::*;
@@ -15,3 +18,6 @@ mod image;
 mod image_handle;
 mod reader;
 mod utils;
+
+#[cfg(test)]
+doc_comment::doctest!("../README.md");